1. address comment. 2. split certificate_types into certificate_lifecycle
