+++ pkg/apis/networking/v1alpha1/certificate_validation.go
@@ -18,7 +18,6 @@ package v1alpha1
 
 import (
 	"github.com/knative/pkg/apis"
-	"k8s.io/apimachinery/pkg/api/equality"
 )
 
 // Validate inspects and validates Certificate object.
@@ -28,10 +27,6 @@ func (c *Certificate) Validate() *apis.FieldError {
 
 // Validate inspects and validates CertificateSpec object.
 func (spec *CertificateSpec) Validate() *apis.FieldError {
-	// Spec must not be empty.
-	if equality.Semantic.DeepEqual(spec, &CertificateSpec{}) {
-		return apis.ErrMissingField(apis.CurrentField)
-	}
 	var all *apis.FieldError
 	// Spec must have at least one DNS Name.
 	if len(spec.DNSNames) == 0 {
@@ -39,7 +34,7 @@ func (spec *CertificateSpec) Validate() *apis.FieldError {
 	} else {
 		for index, dnsName := range spec.DNSNames {
 			if len(dnsName) == 0 {
-				all = all.Also(apis.ErrInvalidArrayValue("DNS Name cannot be empty string.", "dnsNames", index))
+				all = all.Also(apis.ErrInvalidArrayValue("", "dnsNames", index))
 			}
 		}
 	}
